refactor: extract scene base class
