+++ src/client/scene_tutorial.py
@@ -12,7 +12,7 @@ import pygame
 
 from gamecore import config as gconfig
 from gamecore.i18n import gettext as _
-from .app import Scene
+from .scene_base import Scene
 
 
 class TutorialScene(Scene):
